Move function directive to separate file
